feat: add readable and writable memcube-ids
